restructured chats
