+++ utils/getLastTicket.js
@@ -1,0 +1,3 @@
+module.exports = function getLastTicket(chat) {
+  return chat.tickets[chat.tickets.length - 1];
+};
